refactor: refatorado tipagem para simplificar o payload
