+++ src/types/base/base-payload.d.ts
@@ -1,0 +1,6 @@
+export interface BasePayload {
+  action: string
+  pull_request: PullRequest
+  repository: Repository
+  sender: Sender
+}
