+++ src/types/base/href.d.ts
@@ -1,0 +1,3 @@
+export interface Href {
+  href: string
+}
