+++ src/types/base/links.d.ts
@@ -1,0 +1,10 @@
+export interface Links {
+  self: Self
+  html: Html
+  issue: Issue
+  comments: Comments
+  review_comments: ReviewComments
+  review_comment: ReviewComment
+  commits: Commits
+  statuses: Statuses
+}
