+++ src/types/base/reactions.d.ts
@@ -1,0 +1,12 @@
+export interface Reactions {
+  url: string
+  total_count: number
+  '+1': number
+  '-1': number
+  laugh: number
+  hooray: number
+  confused: number
+  heart: number
+  rocket: number
+  eyes: number
+}
