+++ src/types/base/thread.d.ts
@@ -1,0 +1,6 @@
+import { Comment } from '@/types/comment'
+
+export interface Thread {
+  node_id: string
+  comments: Comment[]
+}
