+++ src/types/comment.d.ts
@@ -1,305 +1,13 @@
-export interface User {
-  login: string
-  id: number
-  node_id: string
-  avatar_url: string
-  gravatar_id: string
-  url: string
-  html_url: string
-  followers_url: string
-  following_url: string
-  gists_url: string
-  starred_url: string
-  subscriptions_url: string
-  organizations_url: string
-  repos_url: string
-  events_url: string
-  received_events_url: string
-  type: string
-  site_admin: boolean
-}
-
-export interface Self {
-  href: string
-}
-
-export interface Html {
-  href: string
-}
-
-export interface PullRequest {
-  href: string
-}
-
-export interface Links {
-  self: Self
-  html: Html
-  pull_request: PullRequest
-}
-
-export interface Comment {
-  url: string
-  pull_request_review_id: number
-  id: number
-  node_id: string
-  diff_hunk: string
-  path: string
-  position: number
-  original_position: number
-  commit_id: string
-  original_commit_id: string
-  user: User
-  body: string
-  created_at: Date
-  updated_at: Date
-  html_url: string
-  pull_request_url: string
-  author_association: string
-  _links: Links
-}
-
-export interface Owner {
-  login: string
-  id: number
-  node_id: string
-  avatar_url: string
-  gravatar_id: string
-  url: string
-  html_url: string
-  followers_url: string
-  following_url: string
-  gists_url: string
-  starred_url: string
-  subscriptions_url: string
-  organizations_url: string
-  repos_url: string
-  events_url: string
-  received_events_url: string
-  type: string
-  site_admin: boolean
-}
-
-export interface Repo {
-  id: number
-  node_id: string
-  name: string
-  full_name: string
-  private: boolean
-  owner: Owner
-  html_url: string
-  description?: any
-  fork: boolean
-  url: string
-  forks_url: string
-  keys_url: string
-  collaborators_url: string
-  teams_url: string
-  hooks_url: string
-  issue_events_url: string
-  events_url: string
-  assignees_url: string
-  branches_url: string
-  tags_url: string
-  blobs_url: string
-  git_tags_url: string
-  git_refs_url: string
-  trees_url: string
-  statuses_url: string
-  languages_url: string
-  stargazers_url: string
-  contributors_url: string
-  subscribers_url: string
-  subscription_url: string
-  commits_url: string
-  git_commits_url: string
-  comments_url: string
-  issue_comment_url: string
-  contents_url: string
-  compare_url: string
-  merges_url: string
-  archive_url: string
-  downloads_url: string
-  issues_url: string
-  pulls_url: string
-  milestones_url: string
-  notifications_url: string
-  labels_url: string
-  releases_url: string
-  deployments_url: string
-  created_at: Date
-  updated_at: Date
-  pushed_at: Date
-  git_url: string
-  ssh_url: string
-  clone_url: string
-  svn_url: string
-  homepage?: any
-  size: number
-  stargazers_count: number
-  watchers_count: number
-  language: string
-  has_issues: boolean
-  has_projects: boolean
-  has_downloads: boolean
-  has_wiki: boolean
-  has_pages: boolean
-  forks_count: number
-  mirror_url?: any
-  archived: boolean
-  disabled: boolean
-  open_issues_count: number
-  license?: any
-  forks: number
-  open_issues: number
-  watchers: number
-  default_branch: string
-  allow_squash_merge: boolean
-  allow_merge_commit: boolean
-  allow_rebase_merge: boolean
-  delete_branch_on_merge: boolean
-}
-
-export interface Head {
-  label: string
-  ref: string
-  sha: string
-  user: User
-  repo: Repo
-}
-
-export interface Base {
-  label: string
-  ref: string
-  sha: string
-  user: User
-  repo: Repo
-}
-
-export interface Issue {
-  href: string
-}
-
-export interface Comments {
-  href: string
-}
-
-export interface ReviewComments {
-  href: string
-}
-
-export interface ReviewComment {
-  href: string
-}
-
-export interface Commits {
-  href: string
-}
-
-export interface Statuses {
-  href: string
-}
-
-export interface Repository {
-  id: number
-  node_id: string
-  name: string
-  full_name: string
-  private: boolean
-  owner: Owner
-  html_url: string
-  description?: any
-  fork: boolean
-  url: string
-  forks_url: string
-  keys_url: string
-  collaborators_url: string
-  teams_url: string
-  hooks_url: string
-  issue_events_url: string
-  events_url: string
-  assignees_url: string
-  branches_url: string
-  tags_url: string
-  blobs_url: string
-  git_tags_url: string
-  git_refs_url: string
-  trees_url: string
-  statuses_url: string
-  languages_url: string
-  stargazers_url: string
-  contributors_url: string
-  subscribers_url: string
-  subscription_url: string
-  commits_url: string
-  git_commits_url: string
-  comments_url: string
-  issue_comment_url: string
-  contents_url: string
-  compare_url: string
-  merges_url: string
-  archive_url: string
-  downloads_url: string
-  issues_url: string
-  pulls_url: string
-  milestones_url: string
-  notifications_url: string
-  labels_url: string
-  releases_url: string
-  deployments_url: string
-  created_at: Date
-  updated_at: Date
-  pushed_at: Date
-  git_url: string
-  ssh_url: string
-  clone_url: string
-  svn_url: string
-  homepage?: any
-  size: number
-  stargazers_count: number
-  watchers_count: number
-  language: string
-  has_issues: boolean
-  has_projects: boolean
-  has_downloads: boolean
-  has_wiki: boolean
-  has_pages: boolean
-  forks_count: number
-  mirror_url?: any
-  archived: boolean
-  disabled: boolean
-  open_issues_count: number
-  license?: any
-  forks: number
-  open_issues: number
-  watchers: number
-  default_branch: string
-}
-
-export interface Sender {
-  login: string
-  id: number
-  node_id: string
-  avatar_url: string
-  gravatar_id: string
-  url: string
-  html_url: string
-  followers_url: string
-  following_url: string
-  gists_url: string
-  starred_url: string
-  subscriptions_url: string
-  organizations_url: string
-  repos_url: string
-  events_url: string
-  received_events_url: string
-  type: string
-  site_admin: boolean
-}
+import { Organization } from '@/types/base/oganization'
+import { Thread } from '@/types/base/thread'
+import { User } from '@/types/base/user'
+import { PullRequest, Repository } from '@/types/comment'
 
 export interface CommentPayload {
   action: string
-  comment: Comment
   pull_request: PullRequest
+  thread: Thread
   repository: Repository
-  sender: Sender
+  organization: Organization
+  sender: User
 }
